Add lock to avoid builds running in parallel
When one build is running, we don't want a second one
to run as well as the two builds could step on each
other toes. This change adds a locking mechanism that
waits for the current build to finish before starting
the next build.

+++ tasks/nanoc.rake
@@ -1,34 +1,63 @@
 # frozen_string_literal: true
 
+require "fileutils"
+require "lockf"
 build_dir = Ryo.from(YAML.load_file("./nanoc.yaml")).output_dir
+lockp = File.join Dir.getwd, "tmp", "build.lock"
+FileUtils.touch(lockp)
+lockf = LockFile.new(lockp)
+
 namespace :nanoc do
   task :compile do
+    warn "[build] Acquire lock..."
+    lockf.lock
     ENV["SASS_PATH"] = "./src/css/"
     sh "bundle exec nanoc co"
+  rescue Interrupt
+    warn "SIGINT: exit"
+    exit
+  ensure
+    warn "[build] Release lock..."
+    lockf.release
   end
 
   task :clean do
+    warn "[build] Acquire lock..."
+    lockf.lock
     sh "rm -rf #{build_dir}"
+  ensure
+    warn "[build] Release lock..."
+    lockf.release
   end
 
   task :clean_css do
+    warn "[build] Acquire lock..."
+    lockf.lock
     cssdir = File.join(build_dir, "css")
-    if Dir.exist?(cssdir)
-      sh "rm -rf #{cssdir}"
-    end
+    sh "rm -rf #{cssdir}" if Dir.exist?(cssdir)
+  ensure
+    warn "[build] Release lock..."
+    lockf.release
   end
 
-  task watch: [:compile] do
+  task watch: ['build'] do
+    warn "[build] Acquire lock..."
+    lockf.lock
     require "listen"
     Listen.to File.join(Dir.getwd, "src"), force_polling: true do
       sh "rake build"
     end.start
     sleep
+  rescue Interrupt
+    warn "SIGINT: exit"
+    warn "[build] Release lock..."
+    lockf.release
+    exit
   end
 end
 
 desc "Build the website"
-task build: ["nanoc:clean_css", "nanoc:compile"]
+task build: ["nanoc:compile"]
 
 desc "Trigger a build when src/ is modified"
 task "build:watch" => "nanoc:watch"
